+++ examples/configs/src/view-configs/blin.js
@@ -14,7 +14,7 @@ export const blin2019 = {
       files: [
         {
           fileType: 'image.ome-zarr',
-          url: 'https://minio-dev.openmicroscopy.org/idr/v0.3/idr0062-blin-nuclearsegmentation/6001240.zarr',
+          url: 'https://minio-dev.openmicroscopy.org/idr/v0.4/2022-01-05/idr0062/6001240.zarr',
           options: {
             coordinateTransformations: [
               {