+++ packages/file-types/zarr/src/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 export { default as AnnDataSource } from './AnnDataSource';
 export { default as MuDataSource } from './MuDataSource';
 export { default as ZarrDataSource } from './ZarrDataSource';