+++ packages/utils/other-utils/src/index.js
@@ -1,4 +1,6 @@
+// eslint-disable-next-line react-refresh/only-export-components
 export * from './root';
+// eslint-disable-next-line react-refresh/only-export-components
 export * from './components';
 export {
   getValueTransformFunction,