+++ packages/view-types/feature-list/src/index.js
@@ -1,2 +1,3 @@
-export { FeatureListSubscriber, register } from './FeatureListSubscriber';
+export { FeatureListSubscriber } from './FeatureListSubscriber';
 export { default as FeatureList } from './FeatureList';
+export { register } from './register';