+++ packages/vit-s/src/index.js
@@ -100,4 +100,5 @@ export {
   OptionSelect,
   usePlotOptionsStyles,
 } from './shared-plot-options';
-export * from './schemas';
+// eslint-disable-next-line react-refresh/only-export-components
+export * from './json-schemas';